requirements.txt
fastapi
uvicorn[standard]
python-dotenv
docling
pypdf
llama-index-core
llama-index-readers-docling
llama-index-embeddings-openai
llama-index-node-parser-docling
openai
psycopg[binary]
pgvector
